Added Frankie, a Fortran 90 program for estimating the spectroscopic data (vibrational frequencies, etc.) of a given chemical species.
The vibrational modes are estimated using functional group theory, with
heat capacity data fitting only when needed to fulfill all of the degrees
of freedom. This presents two improvements over THERFIT:

(1) we are using the correct number of vibrational degrees of freedom
rather than using three frequency-degeneracy pairs, and

(2) the estimated frequencies are closer to the actual frequencies because
of the use fo functional group theory.

Currently Frankie estimates the vibrational frequencies and the hindered
rotor frequencies. External rotational modes and symmetry numbers are
not yet available.