[MRH] Updated how RMG handles checking for forbidden structures.
Before: RMG read in the name and graph of the forbidden structures.  When a chemgraph was created, RMG would check this against all stored forbidden structures by using a graph comparison.

Now: RMG reads in the name, graph, # of cabons atoms, # of oxygen atoms, and # of radicals for each forbidden structure.  When a chemgraph is created and RMG is checking if it is a forbidden structure, RMG now checks that the chemgraph of interest has at least the number of carbon atoms, oxygen atoms, and radical number of each individual forbidden structure before performing a graph comparison.  For a system with no oxygen, at least 9 of the 11 current forbidden structures will never be checked via a graph comparison.  For the 1,3-hexadiene case, with pressure dependence turned off, the number of calls to isSub() dropped from ~990,000 to ~45,000.

[MRH] Updated output to user.

Before: RMG woud tell user a new species was created and say "Thermo" and spit out 4 numbers.

Now: RMG does the same thing, but tells the user those 4 numbers are H, S, G, and Cp, respectively, at 298K.