HUGE UPDATE. First in a series of updates designed to make FAME faster and more reliable.
Redesigned the FAME input file to also include a section containing a
list of species. This removes an awkward notation for specifying
spectroscopic data of product channels. A Species type was also created.

The unimolecular and bimolecular isomers are now specified by the Isomer
type. Each Isomer can have one or more Species associated with it.

Added Reaction%kf and Reaction%kb and removed Kij, Fim, Gnj, Jnm.
Similarly, added Isomer%eqDist and removed bi and bn. This improves
speed and readability.

Implemented a new indexing scheme for the ReservoirState mode that
groups by grain, then by well. This improves the banding of the
active-state matrix dramatically, and allows us the use of the banded
LAPACK solver. Also implemented the inversion of the active-state matrix
as a linear solve instead of an inverse. No longer using symmetrization
to do this.

Tentatively fixed bug in collisional transfer probability matrix
normalization by moving the density-of-states ratio so that it is always
less than unity. This appears to only affect the results to within ~10%,
which is plenty good for the crudeness of the collision model.

FAME output reports the k(T, P) values explicitly calculated in addition
to the Chebyshev fit. This allows RMG (and maybe PrIMe) to use bilinear
interpolation instead of the Chebyshev fit.