[MRH] Updated ReactionTemplate.java
Before: The readForbiddenStructure function would look for a file named "fileName" in the rxn template folder.  If it did not find it (which it never would), it would return null.

Now: The function looks for the p_fileName file (which should be forbiddenStructure.txt) in the rxn template folder.

The only rxn family I know of (as of 10Apr2009) that has a forbiddenStructure.txt file is the Birad_recombination family.