[MRH] Updated Thermo.java
Before: Module read in single ChemGraph (no name) and computed the gas-phase + solution-phase chemistry.  These results were output to the screen.

Now: Module reads in multiple ChemGraphs, with names, and computes EITHER the gas-phase or solution-phase chemistry, depending on the "on/off" status of the "Solvation" field.  The outputs are saved to a file named "thermo_output.txt" rather than displayed to the screen.

NOTE: The Species.useInChI will be updated to Species.useSolvation once AJ has committed his solution-phase updates to the CVS.