[MRH] Updated InChI2ChemGraph.java
Before: Read in the InChI string and displayed the Adjacency List to the screen.  Could only read/write one species.

Now: Reads in a list of InChI strings (separated by carriage returns) and outputs a .txt file containing all Adjacency Lists (with the InChI string as the header).  The only display to the user is a "Job successful" message, stating the name of the output file.