[MRH] Created Mol2AdjList
This function converts a .mol file to an RMG adjacency list.  This function was originally created because Linux users were having difficulty getting the InChI2ChemGraph (which should be named InChI2AdjList) to work on a regular basis.  Instead of the user passing the InChI string, the user generates the .mol file separately and then passes that to RMG.  Converting an InChI string to a .mol file has (generally) been much more stable.