[MRH] Updated Species.java
Before: RMG would check the length of each Species chemkin name.  If the length was greater than 16, RMG would rename its chemkin name to SPC(ID).  This check was performed because the Chemkin format for thermodynamic polynomials (NASA format) only allowed 16 characters for the species name.

Now: RMG will update a species' chemkin name if its length is greater than 10.

Chemkin reaction formatting dictates that any "A+B(+m)=C+D(+m)" be of length 52 or less.  Assuming we have 2 reactants and 2 products for a pressure-dependent network, we already use:
	- 2 characters (for the + symbol)
	- 1 character (for the = symbol)
	- 8 characters (4 for each of the (+m) symbols)
This leaves 52 - 11 - 41 characters for 4 species.  Thus, I've changed the criteria from length 16 to length 10.

Considering we rename a species' chemkin name to SPC(ID), this assumes the ID is always less than 100,000 ... which is a safe bet until we mix the memory leak.  ;)