[MRH] Updated PopulateReactions.java
Updated function so that it would recognize/report duplicate kinetics.

Example: H2C*-CH2-CH2-CH3 --> H3C-CH2-*CH-CH3
This can occur via a 3- or 4-member ring; thus, we want to report to the user both sets of kinetics

I define values for the Global.lowTemperature and .highTemperature (these values are completely arbitrary).  They are necessary when the additionalKinetics() call is made but do not affect the reported modified Arrhenius parameters.

I alert the user that a set of duplicates is present by printing the "DUP" tag below each set of duplicate reactions.