[MRH] Created InChI2AdjList.java
This function used to be called InChI2ChemGraph.  However, it was a poor choice of names as the output was not a ChemGraph but a String (in the form of an adjacency list).

The function reads in a .txt file, which contains a list of InChI strings (separated by carriage returns).  The output is a .txt file, containing an adjacency lists for each InChI.