[MRH] Updated ArrheniusKinetics.java
Adding transparency to the RMG user:
Many rxns in the chem.inp file are labeled "!rxn_family_name estimate: ...".

Before: RMG did not report (or keep track) of how it estimated the reported modified Arrhenius coefficients.

Now: When RMG estimates a set of rxn coefficients, it will also report how those numbers were obtained.  For example:

Average of: (node1 && node2 && Average of: (node3 && node4 && node5)).

This will be reported in the chem.inp file for any set of rxn parameters that are reported as "estimate".