[MRH] Updated Structure.java
I've added an additional function: isSpeciesListEquivalentToChemGraphListAsGraphs.

The primary purpose of this function is for checking each new rxn against the Primary Reaction Library.  If the rxn exists in the PRL, RMG will use the rate coefficients parameters stored in the PRL (instead of an estimation routine).

RMG has a similar function isCGListEquivalentAsSpecies.  However, I could not use this function as the primary comparison was between species.  The new rxn (in the form of a structure) contains a list of chemgraphs as its reactants/products because this rxn still sits in the edge (and hence, the species for these chemgraphs have not been made).  However, the rxns in the PRL have already been added to the core and thus the structure contains a list of species for its reactants/products.