[MRH] Created new class SeedMechanism
The class SeedMechanism is replacing the old RMG PrimaryReactionLibrary class.  A PRL now operates similarly to a PrimaryThermoLibrary, in that the reactions in a PRL are no longer automatically added to the core near the beginning of a run.  Instead, RMG checks each new rxn against the rxns stored in the PRL; if the rxn is found, the rate coefficients parameters are grabbed and no estimation technique is required.

Any species and reactions in a SeedMechanism will automatically be thrown into the core at the beginning of the run.  The reactions are added to the list of core rxns and the species, along with any species present in the condition.txt input file, comprise the initial core species.