[MRH] Updated TemplateReactionGenerator.java
After a new species is added to the core, RMG attempts to react it with the already-present core species, against each of the rxn family templates.  Sometimes, this results in RMG not updating the log.txt file for 30-60 minutes (if not longer).  In order to display more information to the user (so they know the program is still running), we know inform the user the species is reacting with the core, against a certain rxn family.