[MRH & RW]: Updated BathGas.java
Before: The order in which colliders was accessed was random (HashMap).  Therefore, if one species and one inert gas existed (CH4 & N2, in the HXD13 test case), the MW would either be that of N2 (if the order accessed by CH4, N2) or ~29.7 (MW_N2 + mf*MW_CH4).  Also, epsilon for an inert gas was defined by getSigma()

Now: The MW, ExpDown, Sigma, Epsilon for the BathGas are weighted sums of all colliders.  Since the sigma/epsilon parameters are not known (in RMG) for He and Ar, we print a message stating so.