[MRH] Updated PopulateReactions.java
Updated how the function writes its output information.  When RMG found a reverse reaction (one that is not given a specific folder in the RMG_database/kinetics directory), it would report the Arrhenius parameters of the forward reaction but display the reverse reaction.

For example, the beta-scission of H2C-(CH2)3-OH = C2H4 + H2CCH2OH would be displayed as C4H9O = C2H4 + C2H5O, but the reported Arrhenius parameters would be that of C2H4 + C2H5O = C4H9O.  In these instances, I reverse the structure of the reaction in the output file.