[MRH] Updated RMG to handle "large" species (i.e. species with more than 4 unique elements).
The GATPFit source code was updated to write the Chemkin/NASA polynomials differently for species containing more than 4 unique elements.  Instead of writing the chemical composition on the first line (between the name and the G Tmin Tmax Tint), it is now written on a new second line; the first line must end in an ampersand (&).  The third-fifth lines are exactly equal to the second-fourth lines from before.

The RMG source code was udpated to look for this new (as well as the old) NASA thermo data formatting.

NOTE: This Chemkin formatting does not work for Chemkin-v.2; it does work for Chemkin-v.4.