[MRH] Updated ReactionTemplate, ReactionModelGenerator, and PopulateReactions
I've restored the "let RMG choose the best kinetics based on the rank and valid TRange" capability.  A ReactionModelGenerator has a new static variable temp4BestKinetics.  This variable is set to the first temperature encountered in the condition.txt file.  When the rxn templates are being read, we get the temp4BestKinetics.

Also, I have RMG check that the name of each species in the condition.txt file does not start with a number.  If it does, we conserve that name in the chem.inp file we generate and Chemkin then complains that a species' name cannot start with a number.