[MRH] Updated PopulateReactions.java
Added an 'else' statement to handle the case where rxn is neither forward nor backward.  Also, added a commented line that would output a reverse rxn as a forward reaction, with its fitted reverse Arrhenius parameters

Example: For an input of HXD13, instead of having the following as ouput:
   CH3 + C5H7 = HXD13 A n E !R_Recombination:
we could output:
   HXD13 = CH3 + C5H7 A' n' E' !Bond_Fission:

I decided to stick with the former as it will allow the user to track down the numbers (since we store the R_Recombination numbers in the kinetics folder).