[MRH] Updated Chemkin.java
Added extra set of parentheses.  Before, if a rxn was reverse and an instance of a TROEReaction, it would be added to the pDepList, along with its forward rxn, resulting in RMG reporting both rxns (forward and reverse) in the chem.inp file, w/o a DUP tag. Furthermore, both rxns were given the same set of Arrhenius parameters.  Running the resulting chem.inp file in Chemkin-v4.1.1 produced an error.