[MRH] Updated PrimaryReactionLibrary.java
Allows multiple kinetics for a given structure within a prl (i.e. A --> B + C has multiple TS and therefore has multiple kinetics).

When RMG reads a PRL, the first instance of a rxn will be added to a temporary LinkedHashSet.  If more than one instance occurs, RMG recognizes that it already exists for this given PRL, so the kinetics of the next instance are added as "additionalKinetics" to the instance already stored in the temporary LinkedHashSet.  Once all rxns in a given PRL have been read in, RMG attempts to add these reactions to the PrimaryReactionLibrary reactionSet.  If there are no instances present, the rxn is added.  If the reactionSet already contains the rxn, the new instance of the rxn is not added.

As an example, suppose a user specifies two PRLs for a given run:

PRL:
Name: PRL-1
Location: ...
Name: PRL-2
Location: ...

Further suppose a rxn A --> B + C has two TS and that both PRLs have modified Arrhenius parameters for both pathways.  The PrimaryReactionLibrary associated with the ReactionModelGenerator will contain the rxn A --> B + C with the kinetics (and additional kinetics) from PRL-1.  If PRL-2 were listed first, the PRL for the RMG would contain the kinetics from PRL-2.