[MRH] Updated rxn/PDepRateConstant, rxn/PDepReaction, rxn/Reaction, rxnSys/Chemkin, and rxnSys/ReactionModelGenerator
Added a new PDepKineticsModel:

I have named the new model RATE.  When the toChemkinString() function is called for a pressure-dependent reaction, both the system temperature and pressure are passed to the function.  If the mode == RATE, the k(T,P) is calculated at the system temperature & pressure.  The reported kinetics for the p-dep rxn in the chem.inp file are:

A + B = C + D k(T,P) 0.0 0.0

Thus, the kinetics as written is only valid for the temperature and pressure given in the condition.txt file.

NOTE: I assume the condition.txt file contains one temperature and one pressure!