[MRH] Updated PresentStatus.java
When PresentStatus is initialized with SystemSnapshot, the inertGas is also copied.

The reason for this change was that the chem.inp file being written did not recognize the InertGas present for the last toChemkinString() call.  We were passing it the PresentStatus, which contained a null InertGas.

Now, the SPECIES portion of the chem.inp file will include any InertGas listed in the condition.txt file