[MRH] Updated calc_freq_code.f90 for CFG
Email from CFG on 3July2009
I took a look at the dat file.
If you look at the dat file, it says that the species has
2 C#CH bonds
2 R-C#C bonds
2 ROOR bonds (which is really 1 as far as frankie is concerned -- more on that later).

where # is a triple bond.
If this is correct, then the species should look like:

HC#COOC#CH

since this species will have 8 atoms and 3 rotors.  Let's pause and consider what a funky peroxide that is.

I have found two bugs in frankie, both related to ROOR, and I have fixed them both.

Here's the first problem:

The current version of frankie has the wrong number of frequencies for ROOR.  ROOR should have two frequencies, not three.  Originally we had three, but I removed one mode.  Instead there should be two frequencies: COO scissor:  350 - 500
OO stretch:  795 - 815
The current source code for frankie has a third frequency:  1185-1215.  I don't recall off the top of my head what that mode was, but in my notes I had removed it.  I thought I had removed it from the source code, but apparently not.  My bad.  This bug by itself wouldn't have caused the code to crash.  What caused the code to crash is the second problem, which is more significant (and much sillier).

Here's the second problem:

Way back in the day, Greg and I noticed that RMG would have a problem double counting ROOR.  To get around this, we decided that the easiest thing to do would be to divide ROOR by 2 from within frankie.  Hence, the dat file shows 2 ROOR, when in fact there is only one (and given the number of atoms, you will quickly see that there could be only one ROOR).

The problem was, stupidly, that the "degeneracy" (i.e. the number of predicted modes) was calculated before I divided by ROOR by 2.  I have now fixed.  My apologies.
So previously, frankie would predict 16 frequencies (as noted by Seyed's output):

2x3 for 2 C#CH
2x2 for 2 C#CR
2x3 for 2 ROOR

for a total of 16,  Naturally, this was a problem, since 3*8-6-3 = 15.

Now we have

2x3 for 2 C#CH
2x2 for 2 C#CR
1x2 for 2 ROOR

for a total of 12.  So now the code will fit three pseudo-frequencies and three hindered rotor potentials.  If I had left ROOR with three frequencies, then we would have predicted 13, rather than 12.  This would be fine, too.  But for the sake of completeness I have reduced the number of frequencies from 3 to 2, since that is what I have in my notes.