[MRH] Added features for Primary Thermo Library
User may now supply RMG w/multiple primary thermo libraries (in the same manner as the primary reaction libraries and seed mechanisms).  The first instance of a species is what RMG will store, so the order of the PTLs in the condition.txt file, as well as the order of the species in the Dictionary.txt file, is important

Also, in the THERMOCHEMISTRY portion of the chem.inp file, a comment is added above each species which comes from a PTL, stating which PTL it came from and the species ID RMG employed.