[MRH] Updated Chemkin.java
For those rxns that have both the pdep and nonPdep rates in the core, this change should have both present in the chem.inp file.

Two iterations ago: Both were present in the chem.inp file, but were listed as duplicates.  According to CHEMKIN, this was an error.

One iteration ago: Only the pdep rxn was present in the chem.inp file (mistakingly).

This iteration: Both will be present in the chem.inp file w/o being labeled duplicates.