[MRH] Updated ArrheniusKinetics.java and ReactionModelGenerator.java
Regarding the chem.inp file:

Previously, when reporting an "estimate", the chem.inp file could contain a long list of "Average of:" statements.  CHEMKIN-v.4.1.1 could not pre-process the particularly long lines properly.

Now, these comments will only appear if user specifies a "Verbose: on" field in the condition.txt file, immediately before InitialStatus.  IF turned off (or not present in the condition.txt file), the user will not receive a verbose comment for the source of the rxn kinetics.