[MRH] Updated ThermoDataEstimator.java
The module was displaying incorrect information for some species (in particular, species with multiple resonance isomers).  This bug was found by Aaron Vandeputte.

In the module, after the species is made, the chemgraph is not necessarily the same (e.g. the species contains resonance isomers, and the adjacency list supplied by the user is not the most stable isomer).  This was causing a discrepancy to be displayed to the screen: the "H=" value we were displaying corresponded to the chemgraph supplied by the user; the ThermoData corresponded to the most stable isomer.

An example of a troublesome chemgraph:
	1 C 0 {2,T}
	2 C 0 {1,T} {3,S}
	3 O 0 {2,S}