[MRH] Updated TemplateReaction.java
Updated RMG source code to accommodate new rxn family by Aaron Vandeputte.  The reaction was labeled with the "thermo_consistence" flag (the rxn template is its own reverse), meaning RMG has to go through extra effort to find Kinetics for the structures (see the hard-coded portions of the code related to H_Abstraction and intra_H_migration).

I include a discussion on the reasoning behind the hard-coding in the java source code.  In the future, we should try to un-hardcode this as a RMG user unfamiliar with Java would find it difficult to construct their own rxn family (if the thermo_consistence flag is used).