[MRH] Updated Graph.java and ReactionAdjList.java
For any rxn family that has 2 products:
RMG distinguished between the 2 products (namely, lumps the "A" and "B") by sorting the graphs by the highestCentralID (e.g. *4).

When implementing one of Aaron Vandeputte's new rxn family templates, he had chemgraphs in his dictionary.txt file that could either have *4 or *5 as the highestCentralID.  This was causing an error: If any of the *5 chemgraphs were interpreted first, the first instance of a *4 chemgraph would throw an error, as the *4 would not match either "A"s or "B"s highestCentralID.

Per the suggestion by Greg Magoon, I have switched the distinguishing between "A"s and "B"s from the highestCentralID to the lowestCentralID.  Everything appears to work for Aaron's new rxn family, as well as the 3 rxn families (the reverse templates of Disproportionation, Disproportionation_O2d, and Cyclic_Ether_Formation) that used to be passed to the altered code in ReactionAdjlist.java