[MRH] Restored restart feature to RMG
chemParser.ChemParser
rxn.ChebyshevPolynomials
rxn.PDepArrheniusKinetics
rxn.PDepRateConstant
rxn.PDepReaction
rxn.Reaction
rxnSys.CoreEdgeReactionModel
rxnSys.ReactionModelGenerator

Restored the "RESTART" feature to RMG.

The core and edge species are added to the CoreEdgeReactionModel before the "InitialStatus" is read from the condition file.  These chemgraphs are sent to a make() method which does not compare the about-to-be-made chemgraph with the SpeciesDictionary.  NOTE: the species must be read after the primaryThermoLibrary, otherwise RMG may estimate the thermo for a species which is present in a primaryThermoLibrary.

The core and edge reactions, along with the p-dep networks, are read in after the contents of the condition.txt file are interpreted.  Although the primaryReactionLibrary and SeedMechanisms have been read in at this point, they have not been added to the CoreEdgeReactionModel.

All species in the Restart files have only one chemgraph stored with it.  RMG will automatically create the resonance structures (and it is relatively cheap to do so).

All reactions in the Restart files are stored in the forward (according to RMG) direction only.  For the edge rxns, it is a bit tricky.  Take HXD13 for example: in the edge, the rxn HXD13=C6H9+H may exist.  RMG would say that C6H9+H=HXD13 is the forward rxn, however, this rxn would not be present in the core or edge (since C6H9 and H are not necessarily in the core).  The trick is to check which side of the rxn has a species that is not contained in the core.  The other direction is the forward direction.

MRH has tested for single & multiple rxn systems, auto and intermediate integration steps, chebyshev and pdeparrhenius modes, completed and early-terminated runs and the results have been promising.  If taking the restart files from a completed run, RMG reads in the files, integrates to the necessary rxn time (or conversion) and does not add additional species/rxns.  If taking the restart files from a prematurely terminated job, RMG will reach the same end state as if the job wasn't prematurely terminated ... most of the time.  MRH has noticed one additional species will be added on occasion, but the overall time of reaction is within the error tolerance.