[MRH] Updated ReactionTemplate.java
There was a problem reading in reactions from a Primary Reaction Library.  RMG labels the forward and reverse reactions with the same kinetics, but identifies the forward reaction (as supplied by the user) using the direction field of the structure (a property of all reactions).  If RMG created that reaction in the mechanism, it first checks whether kinetics for that reaction are given in a Primary Reaction Library.  If found, RMG associates the kinetics with the reaction AND (here was the bug) automatically labeled the reaction as "forward".  Now, RMG also reads the rxn.getStructure.direction field.

For example:
User supplies CH3+CH3 = C2H6 A n E

If user starts with ethane in condition.txt file, one of the initial edge rxns will be C2H6 = CH3 + CH3.  Before, RMG would recognize this reaction to be in a Primary Reaction Library and grab the kinetics (A, n, E).  However, instead of recognizing that the A/n/E are for CH3+CH3=C2H6, RMG associated A/n/E with C2H6=CH3+CH3.  Now, RMG recognizes C2H6=CH3+CH3 as the reverse reaction, and computes the rate for C2H6=CH3+CH3 using the reverse rate (A/n/E) and equilibrium.