[MRH] Updated TemplateReactionGenerator.java
Reduced the amount of garbage collection.  GC will only occur if the free memory is less than 1/3 the total memory.  The "1/3" was arbitrarily chosen by MRH: if the fraction is too close to 1, frequency garbage collection occurs and increases the runtime drastically (as pointed out by Seyed); too close to zero and we risk running out of memory before the next garbage collection.

I have not done a rigorous test to determine the "1/3" value ... just a guess.