[MRH] Updating parseRestartReaction method
I was using .split() method to separate the reactants and products when reading in restart reactions.  However, when attempting to read in an irreversible reaction, and using the .split(["=>']) call, three strings would be returned: the reactants, an empty string, and the products.  I have resolved this by treating the "=" and ">" expressions as normal characters (and not wildcards, potentially).