[MRH] Updated how RMG handles negative concentrations
Before: If a negative concentration occurred, RMG would either:
   a) throw a NegativeConcentrationException, terminating the simulation
   b) check whether the absolute value of the negative concentration was less than 1e-25.  If so, the concentration would be set to zero and the simulation would continue.  If not, a NegativeConcentrationException would be thrown

Now: If a negative concentration occurs, RMG will always check whether the absolute value of that concentration is less than the user-specified Atol (for the ode solver).  If less than Atol, the concentration is set to zero and the simulation continues.  If not, a NegativeConcentrationException is thrown and the program is termianted.