[MRH] Fixed bug in PopulateReactions module
Aaron Vandeputte found a error in the PopulateReactions module: The pre-exponential factor would be incorrect for some reactions; instead of reporting A*(# of unique events), it would simply report A.

MRH also noticed that the same reaction (and kinetics) could appear multiple times in the output file (not incorrect but this can be avoided).

MRH has addressed both issues.  The former was rectified by adding a new method to the TemplateReactionGenerator class.  This new method reacts species against one another for a given reaaction family (instead of all reaction families).  The latter was rectified by creating a new LinkedHashSet (nonDupRxns) and careful bookkeeping.