[MRH] Additional bug fix for ReadRestart option
Before: When reading in the species from the pdepnetworks.txt file, MRH had RMG look up the species in the SpeciesDictionary by chemkin name (as opposed to by chemgraph, to save time).  The bug is that the chemkin name may not be C4H8OJJ(1586) but rather SPC(1586).

Now: If the search using the RMG name (C4H8OJJ(1586)) is null, a search using the proper chemkin name (SPC(1586)) is performed.