[MRH] Updated RMG to address bug in reporting mechanism in chem.inp file (if WriteRestart option was turned on)
Problem: Aaron Vandeputte noticed that some of the pre-exponential factors reported in the mechanism (chem.inp) were the per-event A value, instead of the (per-event A value) * (# of events).

Solution: MRH realized this was occurring if the WriteRestart option was turned on.  What was occurring was the following:

Let's take the standard HXD13 out-of-the-box RMG example, w/pressure dependence turned off.  Furthermore, let's examine the H+CH4=CH3+H2 rxn

RMG would correctly multiply the A factor by the structure's redundancy when calculating the rate to place in the ODEsolver input file.  However, the A reported in the chem.inp file would be the "per event" A.  This was	due to the reaction.toChemkinString() method being called when writing the Restart coreReactions.txt and edgeReactions.txt files.  At the first point of writing the chemkinString for this reaction (when it is still an edge reaction), RMG had not yet computed the redundancy of the structure (as H was not a core	species at the time, but CH3 and H2 were).  When RMG tried to write the chemkinString for the above reaction when writing it to the chem.inp file, using the correct redundancy, the chemkinString already existed and thus the toChemkinString() method was exited immediately.

MRH has replaced the reaction.toChemkinString() call with reaction.toRestartString()	when writing the Restart files, to account for this bug.