Modified compilation scripts in top-level directory; added license.txt.
The Makefile does not actually work yet because the source directory is not
structured in that manner at the moment. The Makefile was designed first, and
the source directory will be reorganized to fit the new Makefile shortly.

The license.txt file contains a copy of the MIT license. Should also add a
README.txt file in the near term.