Rearranged directory structure of Fortran dependencies.
The source/ folder should contain a folder for each module in RMG. Directly
inside that folder is the top-level directory for the source code for that
module; you do not need a "src" directory here, as it is redundant. The
contents of the "run" directory should be placed in examples.

The rearrangement was only done for the dependencies that RMG currently uses.
Deprecated dependencies were left unchanged for now.