Several modifications to .gitignore and the Makefile to enable good workflow.
Added patterns to .gitignore for NetBeans-related project files.

Updated .gitignore to ignore generated RMG output files for the examples.

An exception pattern was added for condition.txt files so that we only save
RMG input files in the examples/ directory.

Removes line used for debugging from JDASSL.java.

Minor modifications to .gitignore and Makefile.

The Makefile does not use debug flags to compile DASSL and DASPK due to the
large volume of code in these that was not written by us. In particular,
DASSL is known to have a bounds check error.

A rule was added to the Makefile for building only the Fortran modules.