Updated .gitignore with new file patterns for NetBeans.
It is now recommended that NetBeans projects be created in a subdirectory
netbeans/ in the top-level directory. This way you can create projects for
each RMG module that you want to develop for, and they won't trip over each
other or the build scripts and Makefiles stored in the repository. The
.gitignore file will automatically ignore anything in the netbeans/ directory.