Added option to SeedMechanism to turn on/off supplemental reaction generation.
When a seed mechanism is used, all seed species and reactions specified are
automatically added to the core. Prior to this commit, RMG would then react
all seed species with each other to generate a complete set of reactions.
This makes some sense, as we won't be able to react the core species with
one another later in the model generation.

However, for large seed mechanisms it may not be practical (or even possible)
to generate all reactions involving the seed species. Furthermore, if you
believe your seed mechanism provides all the important reactions involving
the seed species, you may not want RMG to spend time adding any more.

The implementation of this is a new line GenerateReactions that goes below
the Location line in the SeedMechanism section.