Initialize indices in PDepRateConstant to be zero
The indices t1, t2, p1, p2 were originally initialized to be -1.  This
caused a problem when the temperature of interest (one of the user-
specified temperatures in the TemperatureModel) was also equal to the
lowest temperature for which fame would solve for k(T,P).

The indices are now initalized to be zero, which corresponds to the
lowest pressure and temperature in the list of temperatures and
pressures, respectively.