Bug fix for PDepReactions w/PrimaryReactionLibrary
Before: RMG would read in reaction + kinetics from PrimaryReactionLibrary
and set boolean "fromPrimaryReactionLibrary" to true.  Then, anytime the
algorithm asked for the kinetics, RMG would recognize it to be from a
PrimaryReactionLibrary and thus NOT multiply the pre-exponential factor
by the redundancy of the reaction (as RMG assumes the user supplies the
total "A" factor, rather than the per-event "A" factor).

Problem: When a PDepReaction was created from a reaction, RMG would point
to the high-p-limit kinetics, but not to the "fromPrimaryReactionLibrary"
boolean.  Thus, when RMG asked for the kinetics as it was building the input
file for fame, it would multiply the "A" factor by the reaction's
redundancy.  This would cause the p-dep rate for a given reaction to have
faster kinetics than the high-p-limit!

Solution: When initializing a PDepReaction, the "fromPrimaryReactionLibrary"
boolean for the PDepReaction is set to the value of the Reaction's
"fromPrimaryReactionLibrary".