added code that should (hopefully) fix Franklin's array out of bounds exception issue
Description of the bug: Special thanks to MRH for substantial help in figuring out what was going wrong here. It looks like CO2 was added to the edge, causing it to be added to the species dictionary (no problem so far). Some Troe/Third-Body reactions that Franklin was using contained CO2 as a collider. The collider code only checked whether a species was in the species dictionary before trying to pull its ID from the IDTranslator. When we try to get an ID from the IDTranslator and the value isn't present, the IDTranslator adds the species to the IDTranslator and assigns it an ID. Thus, the size of the IDTranslator and the number of species in the core were different, as CO2 had (incorrectly) been added to the IDTranslator. In getConcentrationFlags, the array is sized by the number of species in the core, but elements are addressed by iterating through the IDTranslator, so when we iterate through the now oversized IDTranslator, we get an array out of bounds exception when we try to access the the n+1-th element of the array.

Description of the fix: Now, with this fix, we also check that the species is in the IDTranslator, so these calls to getRealID should not modify the IDTranslator. There are probably also some other things we could change here (like making a separate function than getRealID which does not modify the IDTranslator or sizing the array in getConcentrationFlags based on the size of IDTranslator rather than the number of core species), but there is a chance this could break something else, so I will just make the IDTranslator check change for now.